pyproject.toml
[tool.poetry]
name = "helmetcheck"
version = "0.1.0"
description = ""
authors = ["Miranda Honkanen <miranda.honkanen@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
